tests/wpt/web-platform-tests/old-tests/submission/Microsoft/dragdrop/dragdrop_005-manual.htm

<!doctype html>
<html>
    <head>
        <title>HTML5 Drag and Drop: Auto state of draggable attribute for div element</title>
        <meta content="text/html; charset=UTF-8" http-equiv="Content-Type"/>
        <link rel="author" title="Microsoft" href="http://www.microsoft.com/"/>
        <link rel="help" href="http://dev.w3.org/html5/spec/dnd.html#the-draggable-attribute"/>
        <meta name="assert" content="Auto state of draggable attribute for div element"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <div id=log></div>
        <div style="border:2px red solid; width:200px; height:50px; display:none;"  id="target">SampleText</div>
        <script type="text/javascript">
            test(function() {assert_false(document.getElementById("target").draggable)}, "Auto state of draggable attribute for div element");
        </script>
    </body>
</html>
